AP MLD: Remove common elements from per STA profile
Even when an element and its contents in the reported link match those
in the reporting link, the element was still added to the per-STA
profile of the reported link. This occurs because the parsed bitmap is
not updated when a match is found, resulting in the element being added
later. This results in per STA profile carrying extra information which
it should not.

Fix this issue by updating the parsed element ID bitmap accordingly.

Fixes: b5359d01ed55 ("AP MLD: Intersect per STA profile with the reporting BSS")
Signed-off-by: Pavithra Ganesan <quic_pavigane@quicinc.com>
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>