tests: Add some more longer duration test cases to long_tests
These test cases seemed to take longer time with UML time-travel and it
is more efficient to run them in the beginning of the test sequence to
avoid leaving a single VM delaying completion of a parallel test run
with large number of VMs. In addition, it looks like some of these test
cases are much more reliable and faster to run at the beginning of a VM
lifetime than at the end.

Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>