mtk: hostapd: change vendor cmd band idx into radio idx
The purpose of this patch is to remove band idx from hostapd.
Regarding the issue of whether we should use radio_idx or
link_id (such as rfeature/wireless ctrl) for per-band
vendor commands, this can be further discussed.
Both of them work, but in my opinion, radio idx is less misleading
and simpler.

Signed-off-by: StanleyYP Wang <StanleyYP.Wang@mediatek.com>