tests: More robust way of killing hung UML VMs
The uml_mconsole halt command may hang when trying to terminate a hung
UML VM, so check for a timeout on that operation and kill the UML
process directly if that happens. In addition, do not try to terminate a
specific VM more than once to avoid confusing debug log entries.

Signed-off-by: Jouni Malinen <j@w1.fi>