OWE: Consider the currently associated transition mode SSID known
When the BSS table size limit is reached the oldest unknown BSS entry is
removed when needing to add a new BSS. This could have resulted in use
of freed memory before wpa_bss_remove_oldest_unknown() was extended to
use wpa_bss_in_use() as a condition for removing a BSS entry. Even with
that issue addressed, it is better to recognize BSS entries that match
the current network profiles SSID in cases where that match is through
the OWE transition mode mechanism. This avoids removing entries that
might be used finding a better BSS.

Signed-off-by: Sunil Ravi <sunilravi@google.com>