P2P2: SSID randomization on group reinvoke
P2P R2 allows the GO to randomize SSID and Group BSSID when a group is
reinvoked. Add support to fetch the group details based on the device
identity block info of the peer with which the group reinvoke is
initiated. As the previous SSID of the network is no longer valid, fetch
the network details based on the ID of the device identity block.

Signed-off-by: Vinay Gannevaram <quic_vganneva@quicinc.com>