NAN USD: Add NAN_UNPAUSE_PUBLISH to cancel pauseState
This allows a publisher to be unpaused when service follow-up has been
completed with a peer. This functionality is not defined in the Wi-Fi
Aware specification for USD, but this is needed to avoid having to wait
for the 60 second pause state expiration since it is not clear when "USD
terminates" without upper layer service specific knowledge.

Signed-off-by: Jouni Malinen <j@w1.fi>