mtk: hostapd: refactor beacon setting after interface setup completion
The original implementation invoked ieee802_11_set_beacons(), which
resulted in excessive and unnecessary beacon settings.

For instance, with 16 BSSes on both the 2.4 GHz and 5 GHz bands, and
additional 16 BSSes on 6 GHz band configured after the 2.4 and 5 GHz
BSSes, calling ieee802_11_set_beacons on the 6 GHz hostapd_iface
triggered beacon settings for each 2.4 GHz and 5 GHz BSS for every 6
GHz BSS. This led to a total of 16 * (1 + 16 + 16) beacon settings,
whereas only one beacon setting per BSS is required to fulfill the
original intent.

This commit refactors the beacon setting process after interface setup
completion to ensuring that each BSS's beacon is set only once.

This change also serves as a fix for commit
2b541601daa (AP: Include an RNR element in Beacon frames for AP MLD).

Signed-off-by: Michael-CY Lee <michael-cy.lee@mediatek.com>