mtk: hostapd: update preference val of neighbor report in BTM req for mld ap
Add support updating the preference values of all BSS neighbor reports
for the MLD AP. Without this commit, hostapd can only update the
preference value of the neighbor report for the current link.

This commit also refactor hostapd_neighbor_set_own_report_pref(). We no
longer need to traverse the neighbor report database; we can simply
traverse the neighbor reports within the BTM request to determine
whether to update the preference value.

mbo_cert=1 is used to determine whether to modify the preference value
of the own BSS report. This check is necessary for handling TGbe 4.2.5.6
and 4.50; the former requires mbo_cert=1 to modify the preference value
of all its neighbor reports, while the latter does not.

Signed-off-by: Howard Hsu <howard-yh.hsu@mediatek.com>