mtk: hostapd: update preference val of neighbor report in BTM req by sta non-preferred channels
Update the preference value of neighbor reports in BTM req by checking
the sta non-preferred channels.

Without this commit, the preference values within the neighbor report
can only be modified based on the STA non-preferred channels if the
neighbor report exists in both neighbor report database and BTM request
simutaneously.

With this commit, we no longer need to traverse the neighbor
report database; we can simply refer to the neighbor reports within the
BTM request to determine whether to modify them based on STA
non-preferred channels.

Signed-off-by: Howard Hsu <howard-yh.hsu@mediatek.com>