mtk: hostapd: fix RNR unsync issue after channel switch
The 6G RNR info in the beacons of the colocated APs (including legacy
2/5G APs and AP-MLDs without a 6G link) is not updated after a channel switch.
The ieee802_11_set_beacon in hostapd_event_ch_switch was removed
in the previous patch to avoid setting the CU flag twice for MLD (csa_after & here).
However, this results in the RNR info in the beacons of those colocated
APs mentioned above not being updated.
Therefore, add ieee802_11_set_beacon_for_colocat to update the beacon
only for the 2/5G legacy AP.

Signed-off-by: StanleyYP Wang <StanleyYP.Wang@mediatek.com>