tests: Fix owe_invalid_assoc_resp with mac80211_hwsim STRICT validation
mac80211 has started rejecting the association attempt based on some
HT/WMM information missing for the (Re)Association Response frame when
in STRICT validation mode and that broke this test case with
mac80211_hwsim due to the Association Response frame being constructed
without HT and WMM information in the test script. Disable HT for now to
avoid this.

Signed-off-by: Jouni Malinen <j@w1.fi>