tests: Update Wireshark filter for MLD test cases
With Wireshark 4.4.0 and above, there are slight changes in the filters for
fetching multi-link control elements and STA profile ID lists. Add support
for these updates to ensure the test cases are compatible with the latest
version of Wireshark.

The changes are:
* Multi-Link Control:
  wlan.eht.multi_link.control instead of wlan.eht.multi_link_control

* STA Profiles LinkIds: It is now Character string.

Latest Documentation:  https://www.wireshark.org/docs/dfref/w/wlan.html

Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>