mtk: hostapd: fix aid allocation
When allocating aid to station, do not check the valid links of stations.
Without this patch, two stations may be allocated to the same aid if
they use different links.

Allocate aid starting from 65. Reserve 64 aids for 16 mbss + 3 * 16 mld
affiliated links.

Signed-off-by: Peter Chiu <chui-hao.chiu@mediatek.com>

The maximum length of interface name is 15.
Without this patch, the AP_VLAN name with default config would be
'phy0.0-ap0.sta65' which exceeds the limitation.

Signed-off-by: Peter Chiu <chui-hao.chiu@mediatek.com>

Free aid bitmap when station is removed.
We add a offset when we get aid from bitmap so we also need to cut
the offset when clean bitmap.

Signed-off-by: Peter Chiu <chui-hao.chiu@mediatek.com>