supplicant: Let user tell us which of multiple radios can be used.
In case of mtk7996 or similar phy that supports multiple underlying
radios, let user specify what bands can be used.  And take this
into account when deciding what can be scanned.  For instance, when
all 3 bands are enabled, and there is already a STA on frequency 6135,
then the scan channels look like this (all of 2.4, all of 5, just the
single channel on 6Ghz).

sta334 (phy #0): scan finished: 6135 2412 2417 2422 2427 2432 2437 2442 2447 2452 2457 2462 5180 5200 5220 5240 5260 5280 5300 5320 5500 5520 5540 5560 5580 5600 5620 5640 5660 5680 5700 5720 5745 5765 5785 5805 5825 5845 5865 5885

Signed-off-by: Ben Greear <greearb@candelatech.com>